pr feedback: move boilerplate to GraphQLUtil and styling
